Fix #22209: NPE in DataSetChangesetBuilder#mergeBounds
This is caused when the latlon of the node is not known, and so
Node#getCoor returned a `null` value.

This patch fixes that by (a) using `isLatLonKnown` to ensure that
the node does have a valid lat/lon, and by not using the Node#getCoor
call (this also avoids an unnecessary allocation).

Signed-off-by: Taylor Smock <tsmock@fb.com>